requirements.txt
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
aiodns==3.4.0
aiofiles==0.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
attrs==25.1.0
audioop-lts==0.2.1
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
Brlapi==0.8.6
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-help-colors==0.9.4
click-plugins==1.1.1.2
colorama==0.4.6
construct==2.8.8
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
custom_gamdl @ git+https://github.com/mediaharbor/custom_gamdl.git@cd57489af774920420baca54a91f441156ecb9a7
custom_streamrip @ git+https://github.com/mediaharbor/custom_streamrip.git@ffe430d709099f26affefee3ac22d158bd6072ad
custom_votify @ git+https://github.com/mediaharbor/custom_votify.git@082900f3fc82118aa75ee102f35b125601801950
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
deezer-py==1.3.6
distro==1.9.0
dnf==4.23.0
fastapi==0.115.12
fedora-third-party==0.10
ffmpy==0.5.0
file-magic==0.4.0
filelock==3.18.0
Flask==3.1.1
fros==1.1
frozenlist==1.6.0
fsspec==2025.3.2
gamdl==2.4.2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.10.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleAPI==0.1.0
googleapis-common-protos==1.69.2
gradio==5.25.0
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.0
idna==3.10
iniconfig==2.1.0
inquirerpy==0.3.4
iso639==0.1.4
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
louis==3.33.0
lxml==5.3.2
m3u8==0.9.0
Mako==1.2.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
mutagen==1.47.0
nftables==0.1
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.47
orjson==3.10.16
packaging==24.2
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.2.3
Paste==3.10.1
pathvalidate==2.5.2
perf==0.1
pexpect==4.9.0
pfzy==0.3.4
pid==2.2.3
pillow==10.4.0
pluggy==1.6.0
ply==3.11
productmd==1.45
progressbar2==4.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
ptyprocess==0.7.0
pwquality==1.4.5
pyapplemusicapi==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase62==1.0.0
pycairo==1.25.1
pycares==4.8.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.23.0
pycryptodomex==3.21.0
pycups==2.0.4
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyenchant==3.2.2
pygit2==1.17.0
Pygments==2.19.1
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pymp4==1.4.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyparted==3.13.0
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.21.2
pytest-mock==3.14.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
python-utils==3.9.0
pytz==2025.2
pyudev==0.24.3
pywidevine==1.8.0
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.15.5-build/kernel-6.15.5/linux-6.15.5-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
re-unplayplay==1.2.9
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
rsa==4.9
ruff==0.11.5
safehttpx==0.1.6
screeninfo==0.8.1
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
semantic-version==2.10.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
shellingham==1.5.4
simple-term-menu==1.6.6
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
starlette==0.46.2
streamrip==2.1.0
systemd-python==235
Tempita==0.5.2
tomlkit==0.7.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.3.0
ursina==7.0.0
uvicorn==0.34.1
waitress==3.0.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xkbregistry==0.3
yarl==1.20.0
yt-dlp==2025.5.22
ytmusicapi==1.10.3
